requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
asttokens
async-timeout==4.0.3
attrs==23.1.0
av==11.0.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
Brotli
bypy==1.8.5
cachetools==5.3.2
causal-conv1d==1.0.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.2.0
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
databricks-cli==0.18.0
datasets==2.15.0
decorator
defusedxml==0.7.1
dill==0.3.7
docker==6.1.3
docopt==0.6.2
easydict==1.11
einops==0.7.0
entrypoints==0.4
exceptiongroup
executing
fastjsonschema==2.20.0
filelock
Flask==3.0.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
fvcore==0.1.5.post20210604
gitdb==4.0.11
GitPython==3.1.40
gmpy2
google-auth==2.26.2
google-auth-oauthlib==1.2.0
greenlet==3.0.2
grpcio==1.60.0
gunicorn==21.2.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.0
iopath==0.1.10
ipython==8.12.3
itsdangerous==2.1.2
jedi
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
lmdb==1.4.1
Mako==1.3.0
mamba-ssm==1.0.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe
matplotlib==3.8.2
matplotlib-inline
mdurl==0.1.2
MedPy==0.4.0
mistune==3.0.2
mkl-fft
mkl-random
mkl-service==2.4.0
ml-collections==0.1.1
mlflow==2.9.1
mmcls==0.25.0
mmcv-full==1.7.0+torch2.1.1cu121
mmdet==2.28.2
mmsegmentation==0.30.0
model-index==0.1.11
monai==1.3.0
mpmath
multidict==6.0.4
multiprocess==0.70.15
MultiScaleDeformableAttention==1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.37
ops==2.12.0
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.1
parso
pavi==0.0.1
pexpect
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.1.0
portalocker==2.0.0
prettytable==3.9.0
progress==1.6
prompt-toolkit
protobuf==4.23.4
ptflops==0.7.3
ptyprocess
pure-eval
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
Pygments
PyJWT==2.8.0
pyparsing==3.1.1
PySocks
python-dateutil==2.8.2
python-hostlist==1.23.0
pytorch-lightning==2.3.0
pytorch-wavelets
pytz==2023.3.post1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.2.0
querystring-parser==1.2.4
referencing==0.35.1
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.4.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
setproctitle==1.2.3
shapely==2.0.2
SimpleITK==2.3.1
simplejson==3.19.3
six
smmap==5.0.1
soupsieve==2.6
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data
sympy
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.4.12
tinycss2==1.3.0
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1
torch_scatter==2.1.2+pt22cu121
torchaudio==2.1.1
torchdiffeq==0.2.3
torchinfo==1.8.0
torchmetrics==1.3.0.post0
torchvision==0.16.1
tornado==6.4.1
tqdm==4.65.2
traitlets
transformers==4.35.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0